Move Query folder to Core.
