+++ BugReport/Reports/Alert.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
-using BugReport.Query;
+using GitHubBugReport.Core.Query;
 
 namespace BugReport.Reports
 {
