+++ BugReport/Reports/Config.cs
@@ -3,9 +3,9 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Xml.Linq;
-using BugReport.Query;
 using BugReport.Util;
 using GitHubBugReport.Core.Issues.Models;
+using GitHubBugReport.Core.Query;
 using GitHubBugReport.Core.Repositories.Models;
 
 namespace BugReport.Reports
