+++ BugReport/Reports/CsvReport/CsvTableReport.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using BugReport.Query;
+using GitHubBugReport.Core.Query;
 
 namespace BugReport.Reports
 {
