+++ BugReport/Reports/ExpressionUntriaged.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using BugReport.Query;
 using BugReport.DataModel;
 using GitHubBugReport.Core.Issues.Extensions;
 using GitHubBugReport.Core.Issues.Models;
+using GitHubBugReport.Core.Query;
 
 namespace BugReport.Reports
 {
