+++ GitHubBugReport.Core/GitHubBugReport.Core.csproj
@@ -56,6 +56,16 @@
     <Compile Include="Issues\Services\IIssueService.cs" />
     <Compile Include="Issues\Services\OctokitIssueService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Query\ExpressionMultiRepo.cs" />
+    <Compile Include="Query\Expressions.cs" />
+    <Compile Include="Query\Expressions_Leaf.cs" />
+    <Compile Include="Query\Expressions_Logical.cs" />
+    <Compile Include="Query\InvalidQueryException.cs" />
+    <Compile Include="Query\Parser.cs" />
+    <Compile Include="Query\RepoExpression.cs" />
+    <Compile Include="Query\SyntaxParser.cs" />
+    <Compile Include="Query\Token.cs" />
+    <Compile Include="Query\Tokenizer.cs" />
     <Compile Include="Repositories\Models\Repository.cs" />
     <Compile Include="Repositories\Services\IRepositoryService.cs" />
     <Compile Include="Repositories\Services\OctoKitRepositoryService.cs" />
