+++ GitHubBugReport.Core/Query/SyntaxParser.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace BugReport.Query
+namespace GitHubBugReport.Core.Query
 {
     internal class QuerySyntaxParser : IDisposable
     {
