+++ GitHubBugReport.Core/Query/Tokenizer.cs
@@ -1,13 +1,8 @@
 using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
-using BugReport.DataModel;
 
-namespace BugReport.Query
+namespace GitHubBugReport.Core.Query
 {
     internal class QueryTokenizer : IDisposable
     {
